reporte terminado
