+++ client/package.json
@@ -27,6 +27,7 @@
     "react-scripts": "^5.0.0",
     "react-sweetalert2": "^0.5.1",
     "react-table": "^7.7.0",
+    "react-to-print": "^2.14.4",
     "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
     "redux-mock-store": "^1.5.4",
