+++ client/src/components/Reporting/ReactTable.css
@@ -1,13 +1,17 @@
   table {
     font-family: Arial, Helvetica, sans-serif;
+    
     border-collapse: collapse;
-    width: 98vw;
-    margin: 1vw;
+    width: 94vw;
+  }
+
+  table tbody{
+    height: 10px;
   }
   
   table td, table th {
-    border: 1px solid #ddd;
-    padding: 8px;
+    border: 2px solid #ddd;
+    padding: 6px;
   }
 
   #paging{
@@ -21,8 +25,8 @@
   table tr:hover {background-color: #ddd;}
   
   table th{
-    padding-top: 12px;
-    padding-bottom: 12px;
+    padding-top: 6px;
+    padding-bottom: 6px;
     text-align: left;
     background-color: #04AA6D;
     color: white;
@@ -30,15 +34,31 @@
 
   tfoot td {
     font-weight: bold;
-    padding-top: 8px;
-    padding-bottom: 8px;
+    padding-top: 6px;
+    padding-bottom: 6px;
     text-align: left;
     background-color: #04AA6D;
     color: white;
+    border-color: #04AA6D;
   }
-
-
 
   input{
     color:black
+  }
+
+  .tableFixHead {
+    overflow-y: auto;
+    height: 78vh;
+    margin: 0.5vw;
+    margin-left: 1vw;
+  }
+
+  
+  .tableFixHead thead th {
+    position: sticky;
+    top: 0;
+  }
+  .tableFixHead tfoot tr {
+    position: sticky;
+    bottom: 0;
   }
