+++ client/src/components/Reporting/utils/ColumnFilter.jsx
@@ -3,11 +3,11 @@ import React from 'react'
 export const ColumnFilter = ({ column }) => {
   const { filterValue, setFilter } = column
   return (
-    <span>
+    <div>
       <input
         value={filterValue || ''}
         onChange={e => setFilter(e.target.value)}
       />
-    </span>
+    </div>
   )
 }
